refactor: create an data file
